Add seeding for listings, also break all unit tests

+++ backend/src/main/java/edu/ntnu/stud/model/base/Listing.java
@@ -53,8 +53,8 @@ public class Listing {
       String description,
       int category,
       int subcategory,
-      long longitude,
-      long latitude,
+      double longitude,
+      double latitude,
       Long ownerId) {
     this.uuid = UUID.randomUUID().toString();
     this.name = name;
